client_app/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
pydantic-settings
tenseal
numpy
concrete-python